+++ torchtitan/experiments/kernels/triton_mg_group_gemm/torchao_pr/mg_grouped_gemm.py
@@ -4,12 +4,12 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-# credit - TMAHelper class, AutoTuning, and flat style indexed forward kernel are derived from FBGemm:
+# credit - flat index forward kernel is derived from FBGemm:
 # https://github.com/pytorch/FBGEMM/blob/main/fbgemm_gpu/experimental/gemm/triton_gemm
 
 # pyre-unsafe
 import functools
-import logging
+
 import os
 import sys
 from typing import Any, Dict, Optional, Tuple
@@ -24,446 +24,12 @@ from triton.runtime import driver  # @manual
 
 sys.path.append(os.path.dirname(os.path.abspath(__file__)))
 
-
-# Configure logging
-logging.basicConfig(
-    level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
+from tma_autotuning import (
+    _NV_CONFIGS,
+    CudaUtils,
+    early_config_prune,
+    TmaDescriptorHelper,
 )
-
-# ===== Supporting utils, CUDA and TMA =====
-
-
-class CudaUtils:
-    @staticmethod
-    def is_cuda() -> bool:
-        """Check if Triton is running on CUDA backend."""
-        return driver.active.get_current_target().backend == "cuda"
-
-    @staticmethod
-    def verify_tma() -> bool:
-        """Check if TMA is supported on the current device."""
-        return (
-            CudaUtils.is_cuda()
-            and torch.cuda.is_available()
-            and torch.cuda.get_device_capability()[0] >= 9
-        )
-
-    @staticmethod
-    def get_num_sms() -> int:
-        """Get the number of streaming multiprocessors on the current device."""
-        if not CudaUtils.is_cuda():
-            raise RuntimeError("Triton is not running on CUDA backend")
-        if not torch.cuda.is_available():
-            raise RuntimeError("CUDA is not available")
-        return torch.cuda.get_device_properties("cuda").multi_processor_count
-
-
-class TmaDescriptorHelper:
-    """Helper class for managing TMA descriptors in Triton kernels."""
-
-    class KernelParamWrapper:
-        """Wrapper to implement the TmaDescKernelParam interface."""
-
-        def __init__(self, desc: torch.Tensor):
-            self.desc = desc
-
-        def tma_desc_cpu_ptr(self) -> int:
-            """Return the CPU pointer to the TMA descriptor."""
-            return self.desc.data_ptr()
-
-    def __init__(self, tma_size: int = 128):
-        """Initialize the TMA descriptor helper.
-
-        Args:
-            tma_size: Size of the TMA descriptor in bytes
-        """
-        if not CudaUtils.verify_tma():
-            raise RuntimeError(
-                "TMA not supported on this device (requires Hopper or newer)"
-            )
-        if "nv_tma_desc_type" not in dir(tl):
-            raise RuntimeError(
-                "TMA grid constant descriptors not supported in your Triton version"
-            )
-
-        self.tma_size = tma_size
-        self.fill_1d_tma_descriptor_inner = driver.active.utils.fill_1d_tma_descriptor
-        self.fill_2d_tma_descriptor_inner = driver.active.utils.fill_2d_tma_descriptor
-        self.descriptors: Dict[str, torch.Tensor] = {}
-
-    def init_tma_descriptor(self, name: str) -> None:
-        """Initialize a TMA descriptor with the given name.
-
-        Call this method outside of the lambda function for grid size.
-        """
-        self.descriptors[name] = torch.empty(
-            self.tma_size, device="cpu", dtype=torch.int8
-        )
-
-    def fill_1d_tma_descriptor(
-        self, name: str, ptr: int, dim: int, block_dim: int, element_size: int
-    ) -> None:
-        """Fill a 1D TMA descriptor.
-
-        Call this method inside the lambda function for grid size.
-        """
-        if name not in self.descriptors:
-            raise ValueError(f"TMA descriptor '{name}' not initialized")
-
-        desc_x = self.descriptors[name]
-        if desc_x.data_ptr() % 64 != 0:
-            raise ValueError("TMA descriptor must be 64-byte aligned")
-        self.fill_1d_tma_descriptor_inner(
-            ptr, dim, block_dim, element_size, desc_x.data_ptr()
-        )
-
-    def fill_2d_tma_descriptor(
-        self,
-        name: str,
-        ptr: int,
-        dim1: int,
-        dim0: int,
-        block_dim1: int,
-        block_dim0: int,
-        element_size: int,
-    ) -> None:
-        """Fill a 2D TMA descriptor.
-
-        Call this method inside the lambda function for grid size.
-        """
-        if name not in self.descriptors:
-            raise ValueError(f"TMA descriptor '{name}' not initialized")
-
-        desc_x = self.descriptors[name]
-        if desc_x.data_ptr() % 64 != 0:
-            raise ValueError("TMA descriptor must be 64-byte aligned")
-        self.fill_2d_tma_descriptor_inner(
-            ptr, dim1, dim0, block_dim1, block_dim0, element_size, desc_x.data_ptr()
-        )
-
-    def get_tma_descriptor_kernel_param(self, name: str) -> KernelParamWrapper:
-        """Get the TMA descriptor kernel parameter for the given name."""
-        if name not in self.descriptors or self.descriptors[name] is None:
-            raise ValueError(f"TMA descriptor '{name}' not initialized")
-        return self.KernelParamWrapper(self.descriptors[name])
-
-
-# ================== End of supporting functions ==================
-
-# ================== Start of FP8 row quantization ==================
-
-
-@triton.autotune(
-    configs=[
-        TConfig({"BLOCK_SIZE": 512}),
-        TConfig({"BLOCK_SIZE": 1024}),
-        TConfig({"BLOCK_SIZE": 2048}),
-        TConfig({"BLOCK_SIZE": 4096}),
-        TConfig({"BLOCK_SIZE": 8192}),
-    ],
-    key=["K"],
-)
-@triton.jit
-def _kernel_quantize_fp8_row(
-    A,
-    A_scale,
-    A_fp8,
-    scale_ub,
-    zero_start_index_M,
-    B,
-    M,
-    N,
-    K,
-    stride_ab,
-    stride_am,
-    stride_an,
-    stride_ak,
-    stride_ob,
-    stride_om,
-    stride_on,
-    stride_ok,
-    stride_zb,  # not used
-    stride_zm,  # not used
-    TL_FP8_DTYPE: tl.constexpr,
-    MAX_FP8: tl.constexpr,
-    EPS: tl.constexpr,
-    CLAMP_MAX: tl.constexpr,
-    BLOCK_SIZE: tl.constexpr,
-    USE_INT64: tl.constexpr,
-) -> None:
-    """Quantize and scale each row.
-
-    Scale per row i is computed as MAX_FP8 / max(abs(A[i, :]))
-
-    Kernel naively iterates through  matrix with [1, BLOCK_SIZE] tiles
-    in a max pass then scale/quantize pass.
-
-    Todo:
-        * Better tiling schemes.
-
-    Args:
-        A (Tensor): higher precision input tensor of 4 dimension.
-        A_scale (Tensor): [B * M * N] reciprocal scale tensor per row.
-        A_fp8 (Tensor): fp8 scaled tensor. A_fp8 = A / a_scale
-        scale_ub (Tensor): [1] Maximum value allowed for scale.
-        B (int): Size of dimenion 0
-        M (int): Size of dimenion 1
-        N (int): Size of dimenion 2
-        K (int): Size of dimenion 3
-        stride_ab (int): Stride of b dimension of A.
-        stride_am (int): Stride of m dimension of A.
-        stride_an (int): Stride of n dimension of A.
-        stride_ak (int): Stride of k dimension of A.
-        stride_ob (int): Stride of b dimension of output.
-        stride_om (int): Stride of m dimension of output.
-        stride_on (int): Stride of n dimension of output.
-        stride_ok (int): Stride of k dimension of output.
-        stride_zb (int): Stride of b dimension of jagged index.
-        stride_zm (int): Stride of m dimension of jagged index.
-        TL_FP8_DTYPE (tl.dtype): Target fp8 datatype.
-        MAX_FP8 (float): Maxmimum expressible value for FP8.
-        EPS (float): Epsilon value for numerical stability.
-        CLAMP_MAX (bool): Whethar to apply scale_ub.
-        Removed:  JAGGED (bool): Whether to use jagged indexing.
-        BLOCK_SIZE (int): Block size for reduction.
-        USE_INT64 (bool): Whether to use int64 indexing for large inputs.
-    """
-    pid = tl.program_id(0)
-    # Use int64 indexing for large inputs. This is slower, but
-    # needed to avoid index overflows.
-    if USE_INT64:
-        pid = pid.to(tl.int64)
-    n_offset = tl.arange(0, BLOCK_SIZE)
-    a_offset_base = (
-        pid // (M * N) * stride_ab
-        + (pid % (M * N)) // N * stride_am
-        + (pid % (M * N)) % N * stride_an
-    )
-    a_fp8_offset_base = (
-        pid // (M * N) * stride_ob
-        + (pid % (M * N)) // N * stride_om
-        + (pid % (M * N)) % N * stride_on
-    )
-
-    K_in = K
-
-    # Calculate max.
-    cur_max = 0.0
-    for _k in range(0, tl.cdiv(K_in, BLOCK_SIZE)):
-        a = tl.load(
-            A + a_offset_base + n_offset * stride_ak,
-            mask=n_offset < K_in,
-            other=0.0,
-        )
-        tile_max = tl.max(tl.abs(a))
-        cur_max = tl.maximum(tile_max, cur_max)
-        n_offset += BLOCK_SIZE
-
-    # Clamp max value appropriately.
-    if CLAMP_MAX:
-        ub = tl.load(scale_ub)
-        cur_max = tl.clamp(cur_max, EPS, ub)
-    else:
-        cur_max = tl.maximum(cur_max, EPS)
-    # Scale and quantize.
-    a_scale = MAX_FP8 / cur_max
-    tl.store(A_scale + pid, 1.0 / a_scale)
-    n_offset = tl.arange(0, BLOCK_SIZE)
-
-    for _k in range(0, tl.cdiv(K, BLOCK_SIZE)):
-        a = tl.load(
-            A + a_offset_base + n_offset * stride_ak,
-            mask=n_offset < K_in,
-            other=0.0,
-        )
-        a_fp8 = a * a_scale
-        # Clamp A to fp8 range to make sure there's no overflow.
-        # This is required for AMD. Nvidia's default saturation
-        # handles it, but it's nice to have anyway.
-        a_fp8 = tl.clamp(a_fp8, -MAX_FP8, MAX_FP8).to(TL_FP8_DTYPE)
-        tl.store(
-            A_fp8 + a_fp8_offset_base + n_offset * stride_ok,
-            a_fp8,
-            mask=n_offset < K,
-        )
-        n_offset += BLOCK_SIZE
-
-
-def triton_quantize_fp8_row(
-    a: torch.Tensor,
-    scale_ub: Optional[torch.Tensor] = None,
-    zero_start_index_M: Optional[torch.Tensor] = None,
-) -> Tuple[torch.Tensor, torch.Tensor]:
-    """
-    Call the triton quantize fp8 row kernel to quantize a tensor to fp8 with row-wise scalings.
-
-    Args:
-        a (Tensor): higher precision input tensor of 4 dimension.
-        scale_ub (Tensor): Maximum allowed value for scale.
-        zero_start_index_M (Tensor): Indicates number of nonzero elements in each row.
-
-    Returns:
-        torch.Tensor: fp8 scaled tensor.
-        torch.Tensor: reciprocal scale tensor per row.
-    """
-    assert a.dim() <= 4, "Triton only supports up to 4 dimension input tensor."
-    a_shape = a.shape
-    while a.dim() < 4:
-        a = a.unsqueeze(0)
-    # if zero_start_index_M is not None:
-    # There should be one value of zero_start_index_M per NxK matrix.
-    #     zero_start_index_M = zero_start_index_M.view(a.shape[0], a.shape[1])
-
-    # Set constant values.
-    pt_fp8_dtype = torch.float8_e4m3fn
-    tl_dtype = tl.float8e4nv
-    max_fp8 = torch.finfo(pt_fp8_dtype).max
-    eps = 1e-12
-
-    num_rows = a.numel() // a.shape[-1]
-    a_scale = torch.empty((num_rows), dtype=torch.float32, device=a.device)
-    a_fp8 = torch.empty(a.shape, device=a.device, dtype=pt_fp8_dtype)
-
-    # If input tensor is sufficiently large, we need to use int64 indexing.
-    use_int64 = a.numel() > (2**31 - 1)
-    grid = (num_rows,)
-
-    _kernel_quantize_fp8_row[grid](
-        a,
-        a_scale,
-        a_fp8,
-        scale_ub,
-        zero_start_index_M,
-        a.shape[0],
-        a.shape[1],
-        a.shape[2],
-        a.shape[3],
-        a.stride(0),
-        a.stride(1),
-        a.stride(2),
-        a.stride(3),
-        a_fp8.stride(0),
-        a_fp8.stride(1),
-        a_fp8.stride(2),
-        a_fp8.stride(3),
-        None,
-        None,
-        # zero_start_index_M.stride(0) if zero_start_index_M is not None else None,
-        # zero_start_index_M.stride(1) if zero_start_index_M is not None else None,
-        TL_FP8_DTYPE=tl_dtype,
-        MAX_FP8=max_fp8,
-        EPS=eps,
-        CLAMP_MAX=scale_ub is not None,
-        # JAGGED=zero_start_index_M is not None,
-        USE_INT64=use_int64,
-    )
-
-    return a_fp8.view(a_shape), a_scale.view(a_shape[:-1])
-
-
-# ================== End of FP8 row quantization ==================
-
-
-# ======  Autotuning utilities ======
-
-_NV_CONFIGS = [
-    triton.Config(
-        {
-            "BLOCK_SIZE_M": block_size_m,
-            "BLOCK_SIZE_N": block_size_n,
-            "BLOCK_SIZE_K": block_size_k,
-        },
-        num_stages=num_stages,
-        num_warps=num_warps,
-        num_ctas=num_ctas,
-    )
-    for block_size_m in [64, 128]
-    for block_size_n in [64, 128, 256]
-    for block_size_k in [64, 128, 256]
-    for num_stages in [3, 4]
-    for num_warps in [4, 8]
-    for num_ctas in [1]
-]
-
-
-def early_config_prune(configs, named_args, dtsize=None, dtype=None, **kwargs):
-    device = torch.cuda.current_device()
-    # Check for all possible pointer parameter names
-    if "grad_input_ptr" in named_args:
-        ptr_name = "grad_input_ptr"
-    elif "c_ptr" in named_args:
-        ptr_name = "c_ptr"
-    elif "grad_weight_ptr" in named_args:
-        ptr_name = "grad_weight_ptr"
-    else:
-        raise KeyError("No recognized pointer parameter found in kernel arguments")
-
-    if dtsize is None:
-        dtsize = named_args[ptr_name].element_size()
-    if dtype is None:
-        dtype = named_args[ptr_name].dtype
-
-    pruned_configs = []
-    for config in configs:
-        kw = config.kwargs
-        BLOCK_M, BLOCK_N, BLOCK_K, num_stages = (
-            kw["BLOCK_SIZE_M"],
-            kw["BLOCK_SIZE_N"],
-            kw["BLOCK_SIZE_K"],
-            config.num_stages,
-        )
-        G, M, N, K = (
-            named_args["G"],
-            named_args["M_BUCKET"],
-            named_args["N"],
-            named_args["K"],
-        )
-
-        # 1. make sure we have enough smem
-        max_shared_memory = driver.active.utils.get_device_properties(device)[
-            "max_shared_mem"
-        ]
-
-        required_shared_memory = (BLOCK_M + BLOCK_N) * BLOCK_K * num_stages * dtsize
-        if required_shared_memory > max_shared_memory:
-            continue
-
-        M_PER_GROUP = M // G
-        MIN_M_TILES = 64
-        # 2. make sure we don't load M tiles that are too big
-        if BLOCK_M > MIN_M_TILES and BLOCK_M > (M_PER_GROUP * 2):
-            continue
-        # 3. make sure we don't load N tiles that are too small
-        if BLOCK_M < 128 and BLOCK_M < (M_PER_GROUP // 2):
-            continue
-
-        num_sm = driver.active.utils.get_device_properties(device)[
-            "multiprocessor_count"
-        ]
-        N_TILES = N // BLOCK_N
-        MIN_N_TILES = 64
-        # 4. make sure we don't load N tiles that are too big
-        if BLOCK_N > MIN_N_TILES and M * N_TILES < num_sm:
-            continue
-        # 5. make sure we don't load N tiles that are too small
-        if BLOCK_N < 128 and M * N_TILES > 2 * num_sm:
-            continue
-        # 6. make sure K can be evenly divided
-        if K % BLOCK_K != 0:
-            continue
-
-        pruned_configs.append(config)
-
-    return pruned_configs
-
-
-# ======== End Autotuning utilities ========
-
-
-# =============== Start Triton Kernels ===============
-"""
-Forward pass for grouped GEMM with Triton, where grouping is M*G
-"""
 
 
 @triton.autotune(
